+++ src/preset/web/providers/config-source.ts
@@ -1,0 +1,14 @@
+import { ConfigSource, createConfigSource } from '../../../config';
+
+/**
+ * Провайдер источника конфигурации.
+ * @return Источник конфигурации.
+ */
+export function provideConfigSource(): ConfigSource {
+  // ВАЖНО: по умолчанию рассчитываем на process.env который предоставляется сборщиком (например webpack)
+  if (typeof process !== 'undefined' && process.env) {
+    return createConfigSource(process.env);
+  }
+
+  return createConfigSource({});
+}
